v0.3.1d: removed strict pixel align pt.1
there is no need to enforce whole interger positions for UI elements. Especially with the recent font changes this was causing alignment problems on high resolution displays.
Elements with non-whole number sizes and positions should be used sparingly, but there's no reason to strictly align them.

+++ src/com/shatteredpixel/shatteredpixeldungeon/ui/RedButton.java
@@ -62,8 +62,8 @@ public class RedButton extends Button {
 		bg.y = y;
 		bg.size( width, height );
 		
-		text.x = x + (int)(width - text.width()) / 2;
-		text.y = y + (int)(height - text.baseLine()) / 2;
+		text.x = x + (width - text.width()) / 2;
+		text.y = y + (height - text.baseLine()) / 2;
 		
 		if (icon != null) {
 			icon.x = x + text.x - icon.width() - 2;
